Server(Back-End)/requirements.txt
Django==1.10.2
djangorestframework==3.4.7
wheel==0.24.0
